+++ src/components/common/dropdownField/DropdownField.styled.ts
@@ -1,0 +1,22 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  margin: 10px 0;
+`;
+
+export const Title = styled.div`
+  font-size: 24px;
+  font-weight: 700;
+  margin: 0;
+`;
+
+export const Subtitle = styled.div`
+  font-size: 13px;
+  font-weight: 300;
+  margin: 10px 0 12px 0;
+  color: #8b95a1;
+`;
+
+export const SelectWrapper = styled.div`
+  margin-top: 8px;
+`;
